refactor: one file per CLI command, flat in src/cli/
src/cli/commands.ts becomes a pure registry (~56 lines): Command type,
imports, array, lookup.

Existing logic files (init, backup, install, tui, import, status) gained
a small *Command export at the bottom — the CLI wrapper lives next to
what it wraps. Commands without a logic file (start, stop, restart,
remove, pair, logs, run, web, proxy) get their own flat file.

No subfolder. No separate command.ts. Shared resolveAgentDir /
serviceOrDaemon helpers in helpers.ts.

Smoke-tested all commands, aliases (ls, status, rm), sub-commands, usage
errors, and unknown-command fallback.
